Allow decoders and encoders to accept schemas as well as types.
The main motivation is to allow `BlockEncoder`s to include non-canonical
schemas in their header. This also makes these streams slightly more
convenient to use for simple use-cases (not requiring a type to be
instantiated separately). The schema parsing logic is the same as for
`avsc.parse`.

Also change name of `schema` argument in all type constructors to
`attrs` to make it more explicit that this argument is different from
the one described above (no JSON decoding or file loading is done on the
former).