Consolidate server with protocol.
For most cases, it wasn't necessary to split the two objects. In cases
where this is desired (e.g. when implementing part of the protocol in
different places), a protocol `copy` method was added (which efficiently
returns an identical protocol, reusing all messages and types, but with
a new set of handlers).