Improve node.js-specific functionality handling.
This lets us now expose the `assemble` function to browserify builds
(the user can provide an import hook which fits the use-case, e.g. from
a text field, using AJAX requests, ...).

This functionality is available in the full and protocol distributions
(both because IDLs are tied to protocol definitions and to keep the
serialization-only distribution as light as possible).