Improve one-way message handling.
One-way message behavior between stateful and stateless transports is
now consistent: listeners are never passed a callback (it was odd to
have the type of transport influence this, breaking separation of
concerns even).

As a side-effect, this change also improves default behavior when no
callback is passed to emitters (even for non-one-way messages). Errors
are now emitted on the corresponding message emitter rather than
directly thrown (where they couldn't be caught).

Also rename "unsupported message" to more explicit "unhandled message"
when a message is sent to a protocol which hasn't implemented a handler
for it.