Move fingerprinting back to types.
This is actually probably worth keeping as-is, even though it somewhat
increases overall binary size. For example when schema inference becomes
available, it will be handy to have a way to disambiguate schemas.

Also allow passing `'null'` to `parse` (requires a bit of extra logic,
but makes it consistent with passing other defined names).