Add option to keep defaults in types' schema.
Defaults are a bit of a special case in schema attributes. They are not
included in the canonical representation of types but are the only other
attribute that can affect whether a schema can parse data (from JSON
encoding) or is compatible with another type. For these reasons, it is
useful to be able to "export" them from a type.