Simplify protocol error handling.
Default error handling is now more intuitive and node-like:

+ Avro's string system errors are by default automatically converted
  to/from actual `Error` objects.
+ Passing `null` as error argument is interpreted as `undefined`
  (meaning the call succeeded).

Since this can lead to ambiguities in rare cases, this behavior can be
turned off with the `strictErrors` option to behave exactly according to
the specification.

Also fix unwrapped and wrapped union handling for logical types and add
optional arguments to `Type.isType`. This makes it much easier to test
whether a type belongs to a set of other types.