Restore enum symbols as array and rename custom long type builder.
The object implementation of symbols was not consistent with other
`getXXXs` methods which all return arrays (`protocol.getMessages`,
`recordType.getFiels`), and lost order. With not static checking, it
isn't clear that there really is a benefit to using symbols as object
fields rather than strings. (Typescript users wouldn't be able to use
this object and would have their own separate anyway.)

Also rename `LongType.using` to `LongType.__with`. This makes it a bit
more explicit that this is a method not meant to be called as part of
the regular flow, but rather to add "core" functionality. But mostly, it
frees up the `use` name for `MessageEmitter` and `MessageListener`
functions, which will provide functionality similar to express' `use`
(avoiding confusion from having completely different implementations
with a `us{e,ing}` name).