Remove references to stateful emitters after EOT.
Stateful emitters and listeners are now fully dereferenced after being
destroyed, making it safer to reuse the underlying streams.

Note that it's likely not worth supporting such behavior for stateless
emitters and listeners because of the overhead of adding the "cleanup"
handlers on each request. Rather, in this case, the streams are assumed
to always be fully owned by the emitter/listener (which is probably fine
since stateless streams should almost always be transient).