Tweak error coercion in non-strict mode.
Error coercion now happens later in the middleware chain (at the very
end), this lets us emit the full stack-trace of error objects on the
server. The default message when emitting "system" errors has also
changed from a generic `INTERNAL_SERVER_ERROR` to the (first line of
the) string. This provides a bit more information to the client by
default (and can be overridden by a custom error formatter if the server
wishes to hide these).