Simplify service error handling.
Channels now never throw errors directly, these are emitted on their
corresponding client or server instead (providing the channel as second
argument). This works well with logging on the server side. On the
client side, the error was relatively redundant with the EOT event,
which will now include an error as second argument if the transmission
ended because of one.

Errors from the auto-generated decoders (in non-object mode) will
likewise destroy their channel with the appropriate error rather than
emit on themselves.