Add more details to typescript annotation file, rearrange it.
Many of the types were generated by (a patched version of) dts-gen.
Because of that, many methods are vague and say `any`.

Rearrange the layout of the declarations to match what dts-gen emitted
(By dts-gen's design, this should match with the signature of `require('avsc')`)

- Keep the precise annotations that were already made, moving some.
- Make interfaces into classes where they correspond to an actual JS
  object that can be used as a class (e.g. has function and function prototype)

A TODO was added to make places using types of `any` specific later.
I'm not familiar enough with this project to be certain of the intent,
and adding the methods/classes is a large enough PR on its own.

The namespaces and classes were generated by `dts-gen`.
Many `any` types are left over, I added a TODO to finish those later

I ran `tsc --strict` in a project that had a dependency of `avsc` to test this.
(The project uses only a few parts of avsc)

- Analysis of avsc.Type.forSchema worked as expected in a project that used
  a patched version of node_modules/avsc, with types/ subfolder added.
  (And toBuffer)