Add type defs for TransportFunction
These were `TODO` so I had a go at adding them as it
was casuing some ugliness when I used avsc from Typescript.

These seem to be correct, and definitely work for me, but it
would be great if someone more familar with the codebase
could check my work :-)

It would have been nice to also constrain the type of
streams further depending on whether the transport would
be read, write, or duplex. I had a quick look at this but
it turned out to be a larger change so I decided to back
off and go with a smaller diff.