Let JS engine copy & initialize buffer in toBuffer
Uint8Array.prototype.slice.call is faster than allocating a buffer then
copying the existing buffer into it. This could be because it's
implemented in native code, because it doesn't need to zero out the
buffer first, or similar reasons.

This gives up to a 30% performance boost on primitive-heavy types.