Remove deflate as a default codec
This is a breaking change. There isn't a good drop-in replacement for
zlib in the browser; browserify-zlib was last updated 6 years ago and
more up-to-date zlib libraries like pako have incompatible API surfaces.

It should be very easy for avsc users to re-add deflate as a codec if
they need it--there's nothing for them to reimplement, and they just
need to pass zlib.[in/de]flateRaw as a codec (as is done in the tests).